% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwsp_model_dw.R
\name{bwsp_model_dw}
\alias{bwsp_model_dw}
\title{Fit Bayesian double Weibull models to time-to-event data}
\usage{
bwsp_model_dw(
  datstan,
  prior.dist = c("fg", "fl", "gg", "ll"),
  chains = 4,
  iter = 11000,
  warmup = 1000
)
}
\arguments{
\item{datstan}{A named list of data for the stanmodel; a data frame can be
reformated with \code{\link{tte2standat}}.}

\item{prior.dist}{A character string indicating the prior distribution for the
parameters of the pgW distribution. Options are
\code{"fg", "fl", "gg", "ll"} (see details).}

\item{chains}{The number of Markov chains to run}

\item{iter}{The total number of iterations per chain (including warmup)}

\item{warmup}{The number of warmup iterations per chain}
}
\value{
List of two stanfit objects; \code{$uncens} contains the stanfit object
obtained from the model to the tte data as is, \code{$cens} the stanfit object
obtained from the model to the tte data censored at mid of observation period.
}
\description{
The function applies the \code{\link[rstan]{sampling}} command to fit two
Weibull models to time-to-event (tte) data with Gamma or Lognormal priors for the
parameters of the Weibull distributions - one to the tte data as is, one to the
tte data censored at mid of observation period.
}
\details{
The function applies the \code{\link[rstan]{sampling}} command to fit two
Weibull models to time-to-event data with Gamma or Lognormal priors for the
parameters of the Weibull distributions - one to the tte data as is, one to the
tte data censored at mid of observation period.

The posterior of each Bayesian model is proportional to the likelihood times the prior.
The likelihood is
\deqn{\mathcal{L}(t| \Theta) = \prod_{i=1}^N S(t_i)^{1-d_i}\cdot f(t_i)^{d_i}}
with \eqn{S(t)} the survival function of the Weibull distribution and \eqn{f(t)} the
density function of the Weibull distribution. The pair \eqn{(t_i, d_i)} are the observed
tte observations.

The priors are either independent univariate Gamma or Lognormal distribution
for the parameters of the pgW distribution.
Implemented distributional choices for the joint prior are products of the following:
\tabular{llll}{
for scales  \tab for shapes  \tab abbreviation \cr
fixed to prior mean \tab Gamma \tab fg \cr
Gamma \tab Gamma \tab ggg \cr
fixed to prior mean \tab Lognormal \tab fl \cr
Lognormal \tab Lognormal \tab ll \cr
}
}
\examples{
# prep the data
head(tte)
standat = tte2priordat_dw(dat = tte,
                          scale.mean = 1,
                          scale.sd = 10,
                          shape.mean = 1,
                          shape.sd = 10,
                          scale_c.mean = 1,
                          scale_c.sd = 10,
                          shape_c.mean = 1,
                          shape_c.sd = 10)
standat

# fit the model
fit = fit_tte_dw(standat,
                 priordist = "ll",
                 chains = 1,
                 iter = 1100,
                 warmup = 100)
                 
# print the summary
fit

}
