% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen_tte.R
\name{datagen_tte}
\alias{datagen_tte}
\title{Data generation}
\usage{
datagen_tte(genpar)
}
\arguments{
\item{genpar}{A vector containing 6 elements, namely:
\enumerate{
\item the sample size N,
\item the background rate br,
\item the ADR rate adr,
\item the ADR mean time quantile percentage,
\item the ADR relative standard deviation from the mean time,
\item the study end time point.
}}
}
\value{
A simulated time-to-event data frame of size N
}
\description{
Simulation of time-to-event (tte) data
}
\details{
Input is a vector with six elements:
\enumerate{
\item the sample size N,
\item the background rate br (observed in population on average),
\item the adverse drug reaction rate as proportion of the background rate,
\item the ADR mean time quantile percentage based on the length of the
observation period,
\item the ADR relative standard deviation from the mean time relative to
the observation period length
\item the study end time point and censoring time for all subjects without
event up to that point.
}

After specification of the input, the data simulation works as follows:

The absolute number of cases due to background causes (other than the adr) is
generated by the binomial distribution with \code{prob = br}.
The absolute number of cases due to the medication is generated by the
binomial distribution with \code{prob = br*adr}. All in all, the expected
number of events/cases within the data set is
\eqn{n \cdot br (1 + adr)}.

The event-times can have values between zero and censor. Time point zero is the date of
the medication (or start of regular medication) and the censor is the
official end time of the cohort study.

For the br cases, the event-times are generated using a uniform distribution on
the interval \link{0, censor}. For the adr cases, event-times follow a normal
distribution. The mean of the normal distribution is at either 0.25, 0.5 or 0.75 times
censor. The standard deviation is defined as
\eqn{ rel.sd \cdot censor}. All generated event-times
\eqn{\leq 0} are set to 1.
}
