% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.calc_auc.R
\name{eval.calc_auc}
\alias{eval.calc_auc}
\title{Calculate AUC of BWSP Test specifications}
\usage{
eval.calc_auc(pc_list)
}
\arguments{
\item{pc_list}{A list containing simulation parameters, test settings, and output paths
(see \code{\link{sim.setup_simpars}} for structure).}
}
\value{
A data frame with one row per ADR-positive simulation scenario, BWSP
test configuration and corresponding AUC value.
}
\description{
Calculates the area under curve (AUC) of the receiver operating characteristic
(ROC) graph with one threshold \insertCite{fawcett2004,lloyd1998}{BWSPsignal}
for all combinations of BWSP test configurations defined in \code{pc_list$test} and
all simulated scenarios.
}
\details{
The function performs BWSP tests across all specified posterior CI types, credibility levels,
and sensitivity options for each simulation run, and calculates AUC using the \code{ROCR} package.
The AUCs are computed for each simulation scenario based on equal numbers of
ADR and no-ADR scenarios.
Scenarios where the number of repetitions deviates from the targeted count
(e.g. due to convergence issues) are returned with \code{NA} AUCs.

The output table provides the base for a ranking of model and test configurations
based on the auc ( see \code{\link{eval.rank_auc}}).
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{bwsp_test}}

\code{\link{eval.rank_auc}}
}
