% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.calc_rope.R
\name{eval.calc_rope}
\alias{eval.calc_rope}
\title{Derive ROPEs from setup parameter combinations in simulation study}
\usage{
eval.calc_rope(rope.infos.row, cred.levels)
}
\description{
Derive ROPEs from setup parameter combinations in simulation study
}
