% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.execution_times.R
\name{eval.execution_times}
\alias{eval.execution_times}
\title{Evaluate execution times by tte and prior distribution}
\usage{
eval.execution_times(pc_list)
}
\arguments{
\item{pc_list}{list containing simulation parameters (see \code{\link{sim.setup_simpars}})}
}
\value{
list with summary statistics (\verb{$summary}), a ggplot2 object (\verb{$plot}),
and cleaned data (\verb{$res.ex_times.table}) which is the base for summary and plot.
}
\description{
Summarizes execution times of the models fitted during the simulation study grouped
by time-to-event (tte) and prior distribution types to guide the
tte and prior distributional choices (along with \code{\link{eval.non_conv_cases}}
and \code{\link{eval.eff_sample_size}}.
}
