% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.merge_rankings.R
\name{eval.merge_rankings}
\alias{eval.merge_rankings}
\title{Evaluation: rank FWSP tests by AUC}
\usage{
eval.merge_rankings(rank_tab_b, rank_tab_f)
}
\description{
Ranks all tried fit & test specifications by AUC.
}
\section{@param rank_tab_b A data frame containing the AUC values for Bayesian test specifications.}{
#' @param rank_tab_f A data frame containing the AUC values for Frequentist test specifications.

#' @return A data frame with the combined rankings of both Bayesian and Frequentist test specifications.
}

