% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.plot_roc.R
\name{eval.plot_roc}
\alias{eval.plot_roc}
\title{Plot ROC Curves for Top Test Specifications}
\usage{
eval.plot_roc(rank_tab, n = 10)
}
\arguments{
\item{rank_tab}{A data frame of ranked test specifications obtained from
\link{`eval.rank_auc_b()`} or \link{`eval.rank_auc_f()`}.}

\item{n}{The number of top-ranked test specifications to plot (default = 10).}
}
\value{
A ggplot object displaying ROC curves with shaded AUC regions.
}
\description{
Plots ROC curves with shaded areas under the curve for the top-ranked test specifications
(based on AUC).
}
\details{
add details and references for ROC curves...
}
\examples{
\dontrun{
ranking_b # output of eval.rank_auc_b()
eval.plot_roc(ranking_b$ranking, n = 3)

ranking_f # output of eval.rank_auc_f()
eval.plot_roc(ranking_b$ranking, n = 3)

# for a comparative study of Bayesian and Frequentist tests 
ranking = dplyr::bind_rows(ranking_b$ranking[, 1:8], 
                           ranking_f$ranking[, 1:5]) \%>\%
  filter(tte.dist == "pgw") \%>\%
  dplyr::arrange(dplyr::desc(AUC))

}

}
