% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pgw_tte.R
\name{fit.fll}
\alias{fit.fll}
\title{fit prespecified pgw stan models}
\usage{
fit.fll(datstan, cores)
}
\description{
The functions essentially just apply the \code{rstan::stan} command with an already prespecified .stan file.
Within a simulation study, the function can be run in parallel on various prior meta-parameter specifications or datasets.
}
