% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.fgg_function.R
\name{fit.ggg}
\alias{fit.ggg}
\title{fit prespecified pgw stan models}
\usage{
fit.ggg(datstan, cores)
}
\description{
The functions essentially just apply the \code{rstan::stan} command with an already prespecified .stan file.
Within a simulation study, the function can be run in parallel on various prior meta-parameter specifications or datasets.
}
