% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.prep_function.R
\name{fit.prep}
\alias{fit.prep}
\title{Prepare Bayesian pgW fitting}
\usage{
fit.prep(survdat, pc)
}
\arguments{
\item{survdat}{time-event-sample}

\item{pc}{one parameter combination used in simulation study}
}
\value{
a list containing all information to be inserted into fit.fgg, fit.fll,
fit.ggg, or fit.lll function.
}
\description{
Contains data preparation for stanmodel.
}
