% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mod_tte.R
\name{fit_mod_tte}
\alias{fit_mod_tte}
\title{Fit stan model to time-to-event data}
\usage{
fit_mod_tte(
  datstan,
  tte.dist = c("w", "dw", "pgw"),
  prior.dist = c("fg", "fl", "gg", "ll"),
  chains = 4,
  iter = 11000,
  warmup = 1000
)
}
\arguments{
\item{datstan}{A named list of data for the stanmodel; a data frame can be
reformated with \code{\link{tte2standat}}.}

\item{tte.dist}{A character string indicating the modelling approach. Options are
\code{"w", "dw", "pgw"}.}

\item{prior.dist}{A character string indicating the prior distribution for the
parameters of the pgW distribution. Options are
\code{"fg", "fl", "gg", "ll"} (see details).}

\item{chains}{The number of Markov chains to run}

\item{iter}{The total number of iterations per chain (including warmup)}

\item{warmup}{The number of warmup iterations per chain}
}
\value{
A stanfit object; in case of \code{tte.dist = "dw"} a list of two stanfit objects
with \code{$uncens} containing the stanfit object
obtained from the model to the tte data as is, and \code{$cens} containing the stanfit object
obtained from the model to the tte data censored at mid of observation period.
}
\description{
Fits a Bayesian model to time-to-event data
for the purpose of performing signal detection tests with \code{\link{bwsp_test}}
based on the posterior samples of the shape parameters of the chosen model
distribution.
}
\details{
The model can be a Weibull,
a double Weibull ( estimating two Weibull models - one to the data as is &
one to the data censored at mid of observation period),
or a power generalized Weibull distribution.

The function applies the \code{\link[rstan]{sampling}} command to fit a Bayesian model to
time-to-event data.
The model can be a Weibull,
a double Weibull (estimating two Weibull models - one to the data as is &
one to the data censored at half of the observation period),
or a power generalized Weibull distribution.

The posterior is proportional to the likelihood times the prior. The likelihood is
\deqn{\mathcal{L}(t| \Theta) = \prod_{i=1}^N S(t_i)^{1-d_i}\cdot f(t_i)^{d_i}}
with \eqn{S(t)} the survival function of the chosen distribution and \eqn{f(t)} the
density function. The pair \eqn{(t_i, d_i)} are the observed
time-to-event observations.

Implemented prior distributions for the scale and shape parameters are products
of the following univariate distributional choices:
\tabular{lll}{
for scale parameter(s) \eqn{\theta} \tab for shape parameter(s) \tab  abbreviation \cr
fixed to prior mean \tab Gamma  \tab fg \cr
Gamma \tab Gamma \tab gg \cr
fixed to prior mean \tab Lognormal \tab fl \cr
Lognormal \tab Lognormal \tab ll \cr
}
}
\examples{
# prep the data
head(tte)
standat = tte2priordat_pgw(dat = tte,
                     tte.dist = "pgw",
                     scale.mean = 1, 
                     scale.sd = 10,
                     shape.mean = 1, 
                     shape.sd = 10,
                     powershape.mean = 1, 
                     powershape.sd = 10)
# fit a pgw model
fit = fit_mod_tte(datstan = standat,  
                  tte.dist = "pgw",         
                  prior.dist = "ll",  
                  chains = 4,
                  iter = 110,          # (be aware that posterior sample
                  warmup = 10)         # is small for demo purpose)
                  
# print the summary
fit



}
