% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pgw_tte.R
\name{fit_pgw_tte}
\alias{fit_pgw_tte}
\title{Fit pgW models with stan}
\usage{
fit_pgw_tte(
  datstan,
  priordist = c("fgg", "fll", "ggg", "lll"),
  chains = 4,
  iter = 11000,
  warmup = 1000
)
}
\arguments{
\item{datstan}{A named list of data for the stan model; a dataframe can be
reformated with \code{tte2standat()}.}

\item{priordist}{A character string indicating the prior distribution for the
parameters of the pgW distribution. Options are
\code{"fgg", "fll", "ggg", "lll"} (see details).}

\item{chains}{The number of Markov chains to run}

\item{iter}{The total number of iterations per chain (including warmup)}

\item{warmup}{The number of warmup iterations per chain}
}
\value{
A stanfit object
}
\description{
The function applies the \code{rstan::sampling} command to fit a pgW model to
time-to-event data with a gamma or lognormal priors for the parameters of the
pgW distribution.
}
\details{
Additional details...
}
