% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tte_pgw.R
\name{fit_tte_pgw}
\alias{fit_tte_pgw}
\title{Fit Bayesian Power generalized Weibull models to time-to-event data}
\usage{
fit_tte_pgw(
  datstan,
  prior.dist = c("fg", "fl", "gg", "ll"),
  chains = 4,
  iter = 11000,
  warmup = 1000
)
}
\arguments{
\item{datstan}{A named list of data for the stanmodel; a tte data frame can be
reformated with \code{\link{tte2priordat_pgw}}.}

\item{chains}{The number of Markov chains to run}

\item{iter}{The total number of iterations per chain (including warmup)}

\item{warmup}{The number of warmup iterations per chain}

\item{priordist}{A character string indicating the prior distribution for the
parameters of the pgW distribution. Options are
\code{"fg", "fl", "gg", "ll"} (see details).}
}
\value{
A stanfit object
}
\description{
The function applies the \code{\link[rstan]{sampling}} command to fit a Power
generalized Weibull (pgw) model to
time-to-event  (tte) data with Gamma or Lognormal priors for the parameters of the
pgW distribution.
}
\details{
The function applies the \code{\link[rstan]{sampling}} command to fit a Bayesian pgw model to
tte data.

The posterior is proportional to the likelihood times the prior. The likelihood is
\deqn{\mathcal{L}(t| \Theta) = \prod_{i=1}^N S(t_i)^{1-d_i}\cdot f(t_i)^{d_i}}
with \eqn{S(t)} the survival function of the pgW distribution and \eqn{f(t)} the
density function of the pgW distribution. The pair \eqn{(t_i, d_i)} are the observed
tte observations.

The priors are either independent univariate Gamma or Lognormal distribution
for the parameters of the pgW distribution.
Implemented distributional choices for the joint prior are products of the following:
\tabular{llll}{
for scale  \tab for shape  \tab for powershape \tab abbreviation \cr
fixed to prior mean \tab Gamma \tab Gamma \tab fg \cr
Gamma \tab Gamma \tab Gamma \tab gg \cr
fixed to prior mean \tab Lognormal \tab Lognormal \tab fl \cr
Lognormal \tab Lognormal \tab Lognormal \tab ll \cr
}
}
\examples{
# prep the data
head(tte)
standat = tte2priordat_pgw(dat = tte,
                           scale.mean = 1, 
                           scale.sd = 10,
                           shape.mean = 1, 
                           shape.sd = 10,
                           powershape.mean = 1, 
                           powershape.sd = 10)
# fit the model
fit = fit_tte_pgw(datstan = standat,  # (be aware that posterior sample
                 prior.dist = "ll",   # is small for demo purpose)
                 chains = 4,
                 iter = 110,
                 warmup = 10)
# print the summary
fit

}
