% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwsp_model.R
\name{fwsp_model}
\alias{fwsp_model}
\title{Fit model to time-to-event data (non-Bayesian)}
\usage{
fwsp_model(dat, tte.dist = c("w", "dw", "pgw"), censor = 365)
}
\arguments{
\item{dat}{A data frame or matrix with time information in first column and
event information (binary status) in second column.}

\item{tte.dist}{A character string specifying the distribution to use for the
model. Options are "w" for Weibull, "dw" for double Weibull, and
"pgw" for power generalized Weibull.}

\item{censor}{time of censoring}
}
\value{
Output of the fitted model. For Weibull models, a \code{summary.survreg} object;
in case of \code{tte.dist = "dw"} a list of two \code{summary.survreg} objects
with \code{$uncens} containing the maximum likelihood estimate
obtained from the model to the tte data as is, and \code{$cens} containing the
maximum likelihood estimate
obtained from the model to the tte data censored at mid of observation period;
for the pgw model, a list containing the maximum likelihood estimate
(log of all parameters) obtained via \code{nlm}.
}
\description{
Fits a time-to-event model using the specified distribution via maximum
likelihood estimation.
}
