% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwsp_test.R
\name{fwsp_test}
\alias{fwsp_test}
\title{Frequentist Shape Parameter Test}
\usage{
fwsp_test(
  mod.output,
  tte.dist = c("w", "dw", "pgw"),
  credlevel = 1 - c(1:10/1000, 2:10/100)
)
}
\arguments{
\item{mod.output}{estimation output resulting from \code{fit_mod_tte_freq(..., tte.dist = "pgw")}}

\item{tte.dist}{character indicating the modelling approach; options are
\code{"w", "dw", "pgw"}}

\item{credlevel}{vector of credibility levels for the tests to be performed}
}
\value{
A vector containing the test results for each credibility level.
}
\description{
Frequentist hypothesis test based on the shape parameter(s) of a distribution of the
Weibull family
}
\details{
This function tests the null hypothesis that the shape parameter(s) of the
Weibull family distribution are equal to one.

The distribution specific definitions of the null hypothesis can be found in
\insertCite{sauzet2024;textual}{BWSPsignal} in Table 1.
}
\references{
\insertAllCited{}
}
