% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwsp_test_dw.R
\name{fwsp_test_dw}
\alias{fwsp_test_dw}
\title{Frequentist double Weibull Shape Parameter Test}
\usage{
fwsp_test_dw(mod.output, credlevel = 1 - c(1:10/1000, 2:10/100))
}
\arguments{
\item{mod.output}{estimation output resulting from \code{fit_mod_tte_freq(..., tte.dist = "w")}}

\item{credlevel}{vector of credibility levels for the tests to be performed}
}
\value{
A vector containing the test results for each credibility level.
}
\description{
Frequentist hypothesis test based on the shape parameter of the Weibull distribution
fitted to the data as is and to the data censored at the middle of the observation period.
}
\details{
The model estimate is a summary of a \code{survival::Survreg} outcome which provides
\eqn{ln(1 / \nu)} as transform of the shape parameter estimate.
The transform \eqn{ln(1 / \nu) = 0} under the null hypothesis \eqn{\nu = 1}.
The shape parameter test is performed on the transform leading the test outcome
equivalent to performing the test based on the shape parameter itself.
}
