% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwsp_test_pgw.R
\name{fwsp_test_pgw}
\alias{fwsp_test_pgw}
\title{Frequentist PgW Shape Parameter Test}
\usage{
fwsp_test_pgw(mod.output, credlevel = 1 - c(1:10/1000, 2:10/100))
}
\arguments{
\item{mod.output}{estimation output resulting from \code{fit_mod_tte_freq(..., tte.dist = "pgw")}}

\item{credlevel}{vector of credibility levels for the tests to be performed}
}
\value{
A vector containing the test results for each credibility level.
}
\description{
Frequentist hypothesis test based on the shape parameters of the power
generalized Weibull distribution.
}
\details{
This function tests the null hypothesis that the logarithm of the shape parameters
of the power generalized Weibull distribution are equal to zero based on the
shape estimates and their estimated standard errors extracted from the estimated
hessian.
}
