% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi_plus_rope.R
\name{hdi_plus_rope}
\alias{hdi_plus_rope}
\title{HDI+ROPE test}
\usage{
hdi_plus_rope(nullregion, credregion)
}
\arguments{
\item{nullregion}{a vector of length two denoting the lower and upper boundary of the null values defined region of practical equivalence (ROPE)}

\item{credregion}{a vector with two entries denoting the lower and upper boundary of the posterior highest density region}
}
\value{
0 if H_0 is accepted, 1 if H_0 is rejected and NA if no decision
}
\description{
Hypothesis test based on the posterior sample of one parameter of interest.
}
\details{
The hypothesis test is based on the comparison of the highest density
interval (HDI) reflecting the posterior distribution of a parameter and
the region of practical equivalence (ROPE) reflecting the null-hypothesis.

Decision rule:
\itemize{
\item If the HDI falls completely inside the ROPE, then accept the null value for practical purposes.
\item If the HDI falls completely outside the ROPE, then reject the null value for practical purposes.
\item Else, withhold a decision.
}

There are many ways to choose an HDI and ROPE depending on the context. For more
information on the concept and exemplary applications see
\insertCite{kruschke2015}{BWSPsignal} and \insertCite{kruschke2018}{BWSPsignal}.
}
\examples{
hdi_plus_rope(nullregion = c(1,2), credregion = c(3,5)) # rope left from hdi and disjunct
hdi_plus_rope(nullregion = c(1,3.5), credregion = c(3,5)) # rope left from hdi, with overlap
hdi_plus_rope(nullregion = c(1,6), credregion = c(3,5)) # rope encompassing hdi
hdi_plus_rope(nullregion = c(6,7), credregion = c(3,5)) # rope right from hdi and disjunct
hdi_plus_rope(nullregion = c(4,7), credregion = c(3,5)) # rope right from hdi with overlap
hdi_plus_rope(nullregion = c(4,4.5), credregion = c(3,5)) # rope encompassed by hdi


}
\references{
\insertRef{kruschke2015}{BWSPsignal}

\insertRef{kruschke2018}{BWSPsignal}
}
\seealso{
\code{\link[HDInterval:hdi]{HDInterval::hdi()}} for calculating highest density intervals from posterior samples.
}
