% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scenario.R
\name{load_scenario}
\alias{load_scenario}
\title{Load single scenario}
\usage{
load_scenario(wd, pc, batchnr = 1)
}
\arguments{
\item{wd}{working directory were ALL the simulation files lie}

\item{pc}{parameter combination (numeric version)}
}
\value{
Returns a table with counts on run simulation scenarios
}
\description{
Load single scenario
}
