% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllk_pgw.R
\name{mllk_pgw}
\alias{mllk_pgw}
\title{Minus-log-Likelihood for pgW distribution}
\usage{
mllk_pgw(par, dat)
}
\arguments{
\item{par}{vector of pgw parameters (scale, shape, powershape)}

\item{dat}{data frame containing tte data}
}
\value{
minus-log-likelihood value given the parameters and the data
}
\description{
Minus-log-likelihood function of the power generalized Weibull distribution
for time-to-event data.
To be used with \code{nlm} or similar optimization functions for maximum likelihood estimation.
}
