% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgw_functions.R
\name{pgw}
\alias{pgw}
\alias{spgw}
\alias{hpgw}
\alias{ppgw}
\alias{dpgw}
\alias{qpgw}
\alias{rpgw}
\title{The power generalized Weibull distribution}
\usage{
spgw(x, scale = 1, shape = 1, powershape = 1, log = FALSE)

hpgw(x, scale = 1, shape = 1, powershape = 1, log = FALSE)

ppgw(x, scale = 1, shape = 1, powershape = 1)

dpgw(x, scale = 1, shape = 1, powershape = 1, log = FALSE)

qpgw(p, scale = 1, shape = 1, powershape = 1)

rpgw(n, scale = 1, shape = 1, powershape = 1)
}
\arguments{
\item{x}{vector of quantiles}

\item{scale}{Scale parameter of the distribution}

\item{shape}{Shape parameter}

\item{powershape}{powershape parameter}

\item{log}{locigal argument; if TRUE, the logarithmized survival probability is returned}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\value{
A vector of cumulative probability values evaluated at each of the inserted quantiles
}
\description{
Survival function, hazard function, cumulative distribution function,
density, quantile function and random generation for the power generalized
Weibull distribution with parameters \code{scale}, \code{shape} and \code{powershape}.
}
\details{
The survival function of the power generalized distribution is (Nikulin and Wu, 2016, p. 15-16):
\deqn{
    S(x) = exp \left\{ 1 - \left[ 1 + \left(\frac{x}{\theta}\right)^{\nu}\right]^{\frac{1}{\gamma}} \right\}.
}
The hazard function is
\deqn{
\frac{\nu}{\gamma\theta^{\nu}}\cdot x^{\nu-1}\cdot  \left[ 1 + \left(\frac{x}{\theta}\right)^{\nu}\right]^{\frac{1}{\gamma-1}}
}
The cumulative distribution function is then \eqn{F(x) = 1 - S(x)} and the density function
is \eqn{S(x)\cdot h(x)}.

If both shape parameters equal 1, the pgw distribution reduces to the exponential distribution
(implemented in \code{d/p/q/rexp} in Rpackage 'stats') with \eqn{\texttt{rate} = 1/\texttt{scale}}
If the powershape parameter equals 1, the pgw distribution simplifies to the Weibull distribution
(implemented in \code{d/p/q/rweibull} in Rpackage 'stats') with the same parametrization.

If parameter values are not specified, they are set as
\code{scale = 1, shape = 1, powershape = 1} per default.
}
