% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgw_tte_gammaprior_scalefixed.R
\name{pgw_fgg}
\alias{pgw_fgg}
\title{Bayesian time-to-event model with PGW likelihood and fix-gam-gam prior}
\usage{
pgw_fgg(tte_datstan, ...)
}
\arguments{
\item{tte_datstan}{A list containing data for the model; can be created using
\code{\link{survdat2pgwstanmodeldat}}}

\item{...}{Additional arguments to pass to \code{\link{rstan::sampling}}}
}
\value{
An object of class \code{stanfit} returned by \code{\link{rstan::sampling}}
}
\description{
Bayesian time-to-event model with PGW likelihood and fix-gam-gam prior
}
\details{
Additional details...
}
