% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pgw.R
\name{plot_pgw}
\alias{plot_pgw}
\title{Visualization of the power generalized Weibull distribution}
\usage{
plot_pgw(scale = 1, shape = 1, powershape = 1, ...)
}
\arguments{
\item{scale}{Scale parameter of the distribution (\eqn{\theta} in formula)}

\item{shape}{Shape parameter (\eqn{\nu} in formula)}

\item{powershape}{Powershape parameter (\eqn{\gamma} in formula)}

\item{...}{optional plot arguments (expept \code{type, lwd, main})}
}
\value{
A plot of the PDF, CDF, hazard and survival function
}
\description{
Visualization of the power generalized Weibull distribution
}
\examples{
plotpgw(scale = 2, shape = 5, powershape = 10)


}
\seealso{
\code{\link{dpgw}}, \code{\link{spgw}}, \code{\link{ppgw}}, \code{\link{hpgw}}.

An interactive version of this plot is available
on \url{https://janoleko.shinyapps.io/pgwd/}.
}
