% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropehdi_function.R
\name{ropehdi}
\alias{ropehdi}
\title{ROPE + HDR test}
\usage{
ropehdi(nullregion, credregion)
}
\arguments{
\item{nullregion}{a vector of length two denoting the lower and upper boundary of the null values defined region of practical equivalence (ROPE)}

\item{scale}{a vector with two entries denoting the lower and upper boundary of the posterior highest density region}
}
\value{
0 if H_0 is accepted, 1 if H_0 is rejected and NA if no decision
}
\description{
ROPE + HDR test
}
\details{
The decision rule is:
\itemize{
\item If the HDI falls completely inside the ROPE, then accept the null value for practical purposes.
\item If the HDI falls completely outside the ROPE, then reject the null value for practical purposes.
\item Else, withhold a decision.
}

For more information see for example

J. K. Kruschke. Rejecting or accepting parameter values in Bayesian estimation. Advances in Methods and Practices in Psychological Science, 1(2):270–280, 2018.

J. K. Kruschke. Doing Bayesian data analysis: a tutorial with R, JAGS, and Stan. Academic Press, 2015.
}
