% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropehdi_combined_extremely_reserved.R
\name{ropehdi_combined_extra_reserved}
\alias{ropehdi_combined_extra_reserved}
\title{Combined ropehdi test (for nu and gamma in one fct)}
\usage{
ropehdi_combined_extra_reserved(credregion.vect, nullregion)
}
\description{
Combined ropehdi test (for nu and gamma in one fct)
}
\details{
Combination mechanism ("intuitive" version) reflects the thought:

If H0 is rejected for both shape parameters nu and gamma of the power generalized
Weibull distribution, then the combined test rejects H0.

All other combinations of single test results indicate no signal.
}
