% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropehdi_combined_reserved.R
\name{ropehdi_combined_reserved}
\alias{ropehdi_combined_reserved}
\title{Combined ropehdi test (for nu and gamma in one fct)}
\usage{
ropehdi_combined_reserved(credregion.vect, nullregion)
}
\description{
Combined ropehdi test (for nu and gamma in one fct)
}
\details{
Combination mechanism ("intuitive" version) reflects the thought:

If H0 is accepted for both shape parameters nu and gamma of the power
generalized Weibull distribution or rejected for one of them and
undecided for the other, then the combined test result is
acceptance of H0. This is equivalent to "no adr signal".

If H0 is rejected for at least one shape parameter or undecided
for both, then the combined test result is rejection of H0.
This is equivalent to a signal.

If
}
