% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.derive_rope.R
\name{sim.derive_rope}
\alias{sim.derive_rope}
\title{Derive ROPEs from setup parameter combinations in simulation study}
\usage{
sim.derive_rope(fit_pc)
}
\description{
Derive ROPEs from setup parameter combinations in simulation study
}
