% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.dist_to_truth_mat_default.R
\name{sim.dist_to_truth_mat_default}
\alias{sim.dist_to_truth_mat_default}
\title{Default specification of distance between adr.when.label and prior.belief}
\usage{
sim.dist_to_truth_mat_default(pc_list)
}
\arguments{
\item{pc_list}{A list containing the simulation parameters obtained from
\link\code{sim.setup_sim_pars}.}
}
\value{
A matrix with rows corresponding to \code{adr.when.label} and columns
corresponding to \code{prior.belief}. The values in the matrix are
\enumerate{
\item 0 if simulated truth and prior.belief are matching,
\item 1 else.
}
}
\description{
This function is used to create the default matrix for the specification of
the distance between the simulated truth about the expected event time
represented by simulation parameter \code{adr.when.label} and \code{prior.belief}
about the expected event time.
}
