% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.fit.to.1.sample.R
\name{sim.fit.to.1.sample}
\alias{sim.fit.to.1.sample}
\title{Simstudy: apply all models to one specific sample}
\usage{
sim.fit.to.1.sample(pc)
}
\arguments{
\item{survdat}{time-to-event sample}
}
\value{
a data frame with 16 rows. Each row contains statistics for one of the
4x4 prior (fll, lll, fgg, ggg) and model (no ADR expected, ADR expected around
1st, 2nd, 3rd quartile of observation period) alternatives.
}
\description{
Conducts all steps from data input to simulation output generation including:
}
\details{
\itemize{
\item data preparation for stanmodel for all prior assumptions
\item fitting all model alternatives
\item extracting and returning relevant statistics from stan output:
\itemize{
\item general information "info" (noch meta), eg. prior mean and sds, fitting specifications etc.
\item information about the posterior distributions of nu and gamma "post"
\item credibility intervals for the tests (will be conducted afterwards)
\item 0 to 100 percentiles of the posterior distribution (to roughly estimate the
probability mass within the region of practical equivalence for the null value)
}
}
}
