% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.load.scenario.R
\name{sim.load.scenario}
\alias{sim.load.scenario}
\title{Load simulated scenario}
\usage{
sim.load.scenario(wd, pc, batchnr = 1)
}
\arguments{
\item{wd}{working directory were ALL the simulation files lie}

\item{pc}{parameter combination (numeric version)}
}
\value{
Returns a table with counts on run simulation scenarios
}
\description{
Load all simulations of a simulated scenario (obtained using \code{\link{sim.rep.1.scenario}}) and return number of run
simulations.
}
