% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.merge.results.R
\name{sim.merge.results}
\alias{sim.merge.results}
\title{merge result table batches from simulationstudy}
\usage{
sim.merge.results(pc_table, wd_load, wd_save, nr_batches = 10, save = F)
}
\arguments{
\item{pc_table}{table of scenarios}

\item{wd_load}{path in which the result batches are stored}

\item{wd_save}{path in which the merged result table will be stored}

\item{nr_batches}{number of batch files per scenario}
}
\value{
one result dataframe containing all simulations (one per row); the
pc specifications are in the first 8 columns, remaining specifications contain
statistics on prior specifications and posterior sample
}
\description{
merge result table batches from simulationstudy
}
