% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.monitor.progress.R
\name{sim.monitor.progress}
\alias{sim.monitor.progress}
\title{Monitor the progress of simulationstudy}
\usage{
sim.monitor.progress(
  pc_table,
  wd = "D:/J/Sciebo/bADR_simstudyres01",
  batch_max = 10
)
}
\arguments{
\item{pc_table}{data frame with the parameter combinations to be used for
the simulation study}

\item{wd}{working directory where the raw result files are stored}

\item{batch_max}{planned number of batches (each batch contains 10 simulation
runs)}
}
\value{
a vector with the number of successfully run simulations for each pc
}
\description{
The function monitors the progress of the simulationstudy by going through
raw result files.
This can be done to find the restarting point when the running of simulations
was interrupted or to investigate which simulations could not be run
successfully (due to convergence issues).
}
