% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.run.R
\name{sim.run}
\alias{sim.run}
\title{run all simulation scenarios}
\usage{
sim.run(pc_list)
}
\arguments{
\item{pc_list}{list of parameter combinations obtained from setup_parameters.R}
}
\description{
run all simulation scenarios
}
