% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.setup_dgp_pars.R
\name{sim.setup_dgp_pars}
\alias{sim.setup_dgp_pars}
\title{Setup simulation scenarios}
\usage{
sim.setup_dgp_pars(N, br, adr.rate, adr.when, adr.relsd, study.period)
}
\arguments{
\item{N}{A scalar or vector of sample sizes.}

\item{br}{A scalar or vector of background rates.}

\item{adr.rate}{A scalar or vector of adverse drug reaction rates.}

\item{adr.when}{A scalar or vector of expected event times (relative number, e.g. 0.5 matches half of study period).}

\item{adr.relsd}{A scalar or vector of relative standard deviations from the adverse drug reaction times.}

\item{study.period}{A scalar specifying the length of the study period.}

\item{adr.when.name}{A vector of description/short name of expected event times (ie. name of simulated truth)}
}
\value{
A data frame with the simulation scenarios that will be inserted into the
\code{\link{datagen_tte}} function.
}
\description{
The function returns the set of simulation scenarios determined by the data generating process (dgp) parameter
values of interest.
}
