% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.setup_fit_pars.R
\name{sim.setup_fit_pars}
\alias{sim.setup_fit_pars}
\title{model fitting parameter specification}
\usage{
sim.setup_fit_pars(
  tte.dist = c("w", "dw", "pgw"),
  prior.belief = c("none", "beginning", "middle", "end"),
  prior.dist = c("fg", "gg", "fl", "ll"),
  list.output = F
)
}
\description{
model fitting parameter specification
}
