% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.setup_simpars.R
\name{sim.setup_sim_pars}
\alias{sim.setup_sim_pars}
\title{Set up simulation parameters}
\usage{
sim.setup_sim_pars(
  N,
  br,
  adr.rate,
  adr.when,
  adr.when.label,
  adr.relsd,
  study.period,
  tte.dist,
  prior.dist,
  post.ci.type,
  cred.level,
  sensitivity.option,
  reps = 100,
  batch.size = 10,
  batch.nr = reps/batch.size,
  resultpath = paste0(getwd(), "/results_raw"),
  stanmod.chains = 4,
  stanmod.iter = 11000,
  stanmod.warmup = 1000
)
}
\arguments{
\item{N}{A scalar or vector of sample sizes.}

\item{br}{A scalar or vector of background rates.}

\item{adr.rate}{A scalar or vector of adverse drug reaction rates.}

\item{adr.when}{A scalar or vector of expected event times (relative number, e.g. 0.5 matches half of study period).}

\item{adr.relsd}{A scalar or vector of relative standard deviations from the adverse drug reaction times.}

\item{study.period}{A scalar specifying the length of the study period.}

\item{tte.dist}{A character string indicating the modelling approach. Options are
\code{"w", "dw", "pgw"} (see also \link\code{fit_mod_tte}).}

\item{prior.dist}{A character string indicating the prior distribution for the
parameters of the pgW distribution. Options are
\code{"fg", "fl", "gg", "ll"} (see also \link\code{fit_mod_tte}).}

\item{post.ci.type}{A character string indicating whether to extract equal tailed
intervals (\code{"ETI"}) or highest posterior density intervals (\code{HDI}) as
credibilty interval/region for BWSP testing (see \link\code{bwsp_test}).}

\item{cred.level}{A scalar or vector of credibility levels to be tried for construction
of region of practical equivalence (ROPE) and posterior credibility interval (CI).}

\item{sensitivity.option}{A scalar or vector of sensitivity options to be tried for the BWSP test.}

\item{reps}{The number of repetitions for each simulation scenario.}

\item{batch.size}{The number of simulation repetitions to be saved in a batch}

\item{batch.nr}{per default \code{reps/batch.size}; the number of batch files
in resultpath.}

\item{resultpath}{The directory where intermediate results of the simulation
are saved.}

\item{stanmod.chains}{The number of Markov chains to run.}

\item{stanmod.iter}{The total number of iterations per chain (including warmup).}

\item{adr.when.name}{A vector of description/short name of expected event times (ie. name of simulated truth)}

\item{standmod.warmup}{The number of warmup iterations per chain.}
}
\description{
With one partly interactive function
data generation, model fitting, number of simulation runs and evaluation are
set up and gathered in a parameter combination (pc) list, from which the
information are called in following steps of a simulation study.
}
\details{
The tuning parameters are used to evaluate the performance of different BWSP
signal detection tests applied to data scenarios of interest.
The set of tuning parameters consisting of the choice of the
\enumerate{
\item distribution chosen for the tte model, must be a subset out of "w", "dw"
and pgw",
\item prior belief about an adverse event being and adverse drug reaction, and if
so, when it is expected to occur
(this will trigger promts asking for prior means and standard deviations for
all parameters reflecting that belief; have a look at our vignette about
prior elicitation to formalize the prior belief),
prior distribution chosen for scale and shape parameters, must be a subset
out of "fgg", "ggg", "fll" and "lll",
\item type of posterior credible interval, must be a subset out of "ETI"
(equal-tailed interval) and "HDI" (highest density interval).
}

Batch saving is done to prevent losing simulation results in case of an
interuption of simulation.
}
