% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.setup_test_pars.R
\name{sim.setup_test_pars}
\alias{sim.setup_test_pars}
\title{setup of test parameters}
\usage{
sim.setup_test_pars(
  post.ci.type = c("ETI", "HDI"),
  cred.level = seq(0.5, 0.95, by = 0.05),
  sensitivity.option = 1:3
)
}
\description{
setup of test parameters
}
