% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.stanfit.to.fitstats.R
\name{sim.stanfit.to.fitstats}
\alias{sim.stanfit.to.fitstats}
\title{extract meta and prior statistics from stanfit object 2.0}
\usage{
sim.stanfit.to.fitstats(stanfit.object, stan.dat)
}
\arguments{
\item{stanfit.object}{the estimated stan model output}

\item{stan.dat}{the data inserted for model fitting}
}
\value{
Information about the model fitting, such as:
\item{run.min}{running time in minutes}
\item{th.pr.mean}{prior mean of parameter theta}
\item{th.pr.sd}{prior standard deviation of parameter theta}
\item{nu.pr.mean}{prior mean of parameter nu}
\item{nu.pr.sd}{prior standard deviation of parameter nu}
\item{ga.pr.mean}{prior mean of parameter gamma}
\item{ga.pr.sd}{prior standard deviation of parameter gamma}
}
\description{
The function extracts information about the model fitting and prior specifications for the parameters to be estimated.
}
