% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte2priordat_dw.R
\name{tte2priordat_dw}
\alias{tte2priordat_dw}
\title{Prior and data setup for Bayesian double Weibull model fitting}
\usage{
tte2priordat_dw(
  dat,
  scale.mean,
  scale.sd,
  shape.mean,
  shape.sd,
  scale_c.mean,
  scale_c.sd,
  shape_c.mean,
  shape_c.sd
)
}
\arguments{
\item{dat}{Matrix or data frame with time in the first column and event status in the second column.}

\item{scale.mean}{The a priori expected mean of the scale parameter.}

\item{scale.sd}{The priori expected standard deviation of the scale parameter.}

\item{shape.mean}{The priori expected mean of the shape parameter.}

\item{shape.sd}{The priori expected standard deviation of the shape parameter.}

\item{scale_c.mean}{The a priori expected mean of the scale parameter for the
data censored at half of the observation period.}

\item{scale_c.sd}{The priori expected standard deviation of the scale parameter
for the data censored at half of the observation period.}

\item{shape_c.mean}{The priori expected mean of the shape parameter for the
data censored at half of the observation period.}

\item{shape_c.sd}{The priori expected standard deviation of the shape parameter
for the data censored at half of the observation period.}
}
\description{
Setup of prior specifications and time-to-event
data set as a list suitable for model fitting with \code{\link{bwsp_model}}.
}
\details{
The function converts time-to-event data and prior specifications for location ('.mean') and
precision ('.sd') of the double Weibull parameters
to a suitable format for \code{rstan} model fitting
which is performed within \code{\link{bwsp_model}}.

Prior means suitable to reflect the prior belief can be found by plotting the
hazard and estimating the expected event time under different parameter combinations
using \code{\link{plot_pgw}} with 'powershape = 1' (reducing the pgw distribution
to the Weibull distribution) or \url{https://janoleko.shinyapps.io/pgwd/}.

Prior standard deviations should reflect the uncertainty about the prior belief
(i.e. set smaller standard deviation in case of high certainty about prior belief vs. larger
standard deviation in case of low certainty).
}
