% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte2priordat_w.R
\name{tte2priordat_w}
\alias{tte2priordat_w}
\title{Prior and data setup for Bayesian Weibull model fitting}
\usage{
tte2priordat_w(dat, scale.mean, scale.sd, shape.mean, shape.sd)
}
\arguments{
\item{dat}{Matrix or data frame with time in the first column and event status in the second column.}

\item{scale.mean}{The a priori expected mean of the scale parameter.}

\item{scale.sd}{The priori expected standard deviation of the scale parameter.}

\item{shape.mean}{The priori expected mean of the shape parameter.}

\item{shape.sd}{The priori expected standard deviation of the shape parameter.}
}
\description{
Setup of prior specifications and reformated time-to-event
data set as a list suitable for model fitting with \code{\link{fit_w_tte}}.
}
\details{
The function converts time-to-event data and prior specifications for location ('.mean') and
precision ('.sd') of the Weibull parameters
to a suitable format for \code{rstan} model fitting
which is performed within \code{\link{fit_w_tte}}.

Prior means suitable to reflect the prior belief can be found by plotting the
hazard and estimating the expected event time under different parameter combinations
using \code{\link{plot_pgw}} with 'powershape = 1' (reducing the pgw distribution
to the Weibull distribution) or \url{https://janoleko.shinyapps.io/pgwd/}.

Prior standard deviations should reflect the uncertainty about the prior belief
(i.e. set smaller standard deviation in case of high certainty about prior belief vs. larger
standard deviation in case of low certainty).
}
