% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte2standat.R
\name{tte2standat}
\alias{tte2standat}
\title{reformat time-to-event data for usage in \code{fit_pgw_tte()}}
\usage{
tte2standat(
  dat,
  scale.mean,
  scale.sd,
  shape.mean,
  shape.sd,
  powershape.mean,
  powershape.sd
)
}
\description{
Reformatting of a matrix or data frame containing a time-to-event
data set with time in the first column and the event status in the second column.
}
\details{
Converts time-event data to a suitable format for \code{rstan} model fitting.
}
