% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSPsignal-package.R
\docType{package}
\name{WSPsignal-package}
\alias{WSPsignal-package}
\alias{WSPsignal}
\title{The 'WSPsignal' package.}
\description{
An R package to perform Bayesian Weibull Shape Parameter (BWSP)
tests for signal detection.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan.
R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Julia Dyck \email{j.dyck@uni-bielefeld.de}

Authors:
\itemize{
  \item Odile Sauzet
}

}
