% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.non_conv_cases.R
\name{eval.non_conv_cases}
\alias{eval.non_conv_cases}
\title{Based on simulation study batch files the function groups the data by \code{tte.dist}
and \code{prior.dist}, and calculates the total planned and total not run repetitions,
and the proportion of repetitions that were not run
to guide the
tte and prior distributional choices (along with \code{\link{eval.execution_times}}
and \code{\link{eval.eff_sample_size}}}
\usage{
eval.non_conv_cases(pc_list)
}
\arguments{
\item{pc_list}{list containing simulation parameters (see \code{\link{sim.setup_simpars}})}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{tte.dist}: The time-to-event distribution.
\item \code{prior.dist}: The prior distribution.
\item \code{total.notrun}: The total number of repetitions that were not run.
\item \code{total.planned}: The total number of planned repetitions.
\item \code{prop.notrun}: The proportion of repetitions that were not run.
}
}
\description{
Based on simulation study batch files the function groups the data by \code{tte.dist}
and \code{prior.dist}, and calculates the total planned and total not run repetitions,
and the proportion of repetitions that were not run
to guide the
tte and prior distributional choices (along with \code{\link{eval.execution_times}}
and \code{\link{eval.eff_sample_size}}
}
