% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.non_conv_cases.R
\name{eval.non_conv_cases}
\alias{eval.non_conv_cases}
\title{This function processes a list containing simulation data and computes
the total number of repetitions that were completed and not completed
based on the progress of the simulation. It groups the data by \code{tte.dist}
and \code{prior.dist}, calculates the total planned and total not run repetitions,
and computes the proportion of repetitions that were not run.}
\usage{
eval.non_conv_cases(pc_list)
}
\arguments{
\item{pc_list}{A list generated with \code{\link{sim.setup_simpars}} after
simulations were run.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{tte.dist}: The time-to-event distribution.
\item \code{prior.dist}: The prior distribution.
\item \code{total.notrun}: The total number of repetitions that were not run.
\item \code{total.planned}: The total number of planned repetitions.
\item \code{prop.notrun}: The proportion of repetitions that were not run.
}
}
\description{
This function processes a list containing simulation data and computes
the total number of repetitions that were completed and not completed
based on the progress of the simulation. It groups the data by \code{tte.dist}
and \code{prior.dist}, calculates the total planned and total not run repetitions,
and computes the proportion of repetitions that were not run.
}
