% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.rank_auc.R
\name{eval.rank_auc}
\alias{eval.rank_auc}
\title{Evaluation: rank BWSP tests by AUC}
\usage{
eval.rank_auc(auc.df)
}
\arguments{
\item{auc.df}{A data frame containing AUC results returned by \link{eval.calc_auc}.}
}
\value{
A list containing
\itemize{
\item \code{$ranking}: Ranking of fit and BWSP test specifications according
to AUC averaged over all sample scenarios given a correct specification of prior belief.
\item \code{$effect.of.N}: Effect of sample size on AUC for the optimal fit
and BWSP test specification given a correct specification of prior belief.
\item \code{$effect.of.br}: Effect of background rate on AUC for the optimal fit
and BWSP test specification given a correct specification of prior belief.
\item \code{$effect.of.adr.rate}: Effect of ADR rate on AUC for the optimal fit
and BWSP test specification given a correct specification of prior belief.
\item \code{$effect.of.adr.when}: Effect of true expected event times on AUC
for the optimal fit and BWSP test specification given a correct specification of prior belief.
\item \code{$effect.of.adr.relsd}: Effect of relative standard deviation of event
time on AUC for the optimal fit and BWSP test specification given a correct
specification of prior belief.
\item \code{$effect.of.dist.prior.to.truth}: Effect of distance of prior belief
to true adr.when on AUC for the optimal fit and BWSP test specification given a
correct specification of prior belief.
}
}
\description{
Ranks all tried fit & test specifications by AUC.
}
\examples{
\dontrun{
# Filtering for a specific tte and prior distribution (for instance based on
# evaluation of execution times, convergence and 
 auc_dw_ll <- dplyr::filter(aucs, tte.dist == "dw", prior.dist == "ll")
 eval.rank_auc(auc_dw_ll)
}
}
