% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.prep_function.R
\name{fit.prep}
\alias{fit.prep}
\title{Prepare Bayesian pgW fitting}
\usage{
fit.prep(ttedat, pc)
}
\arguments{
\item{ttedat}{time-to-event data set}

\item{pc}{vector representing one parameter combination used in simulation study}
}
\value{
list containing all information to be inserted as \code{datstan}
argument into the \code{\link{fit_pgw_tte} function.
}
\description{
Formatting of simulated data generated with \code{link{datagen_tte}} as building
block for tuning the Bayesian WSP test via simulation study.
}
