% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pgw_tte.R
\name{fit_pgw_tte}
\alias{fit_pgw_tte}
\title{Fit pgW models with stan}
\usage{
fit_pgw_tte(
  datstan,
  priordist = c("fgg", "fll", "ggg", "lll"),
  chains = 4,
  iter = 11000,
  warmup = 1000
)
}
\description{
The function apply the \code{rstan::stan} command with an already prespecified .stan file.
Within a simulation study, the function can be run in parallel on various prior meta-parameter specifications or datasets.
}
