% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamprior_repar.R
\name{gamprior_repar}
\alias{gamprior_repar}
\title{Reparametrization of Gamma mean and sd}
\usage{
gamprior_repar(m_t, s_t)
}
\arguments{
\item{m_t}{mean of the Gamma distribution}

\item{s_t}{standard deviation of the Gamma distribution}
}
\value{
a vector with the \code{shape}
and \code{rate} of the Gamma distribution
}
\description{
\code{gamprior_repar} can be used after specifying a Gamma prior in terms of
mean \code{m_t} and standard deviation \code{s_t} to calculate the \code{shape}
and \code{rate} as defined and used in \code{\link[stats]{dgamma}}.
}
\details{
Expectation (mean) and standard deviation of a Gamma distributed random variable \eqn{X}
parametrized in terms of the shape parameter $\alpha$
and rate parameter $\beta$ are given by

\deqn{E(X) = \frac{\alpha}{\beta}}
\deqn{sd(X) = \frac{\alpha}{\beta^2}}

For \eqn{E(X), \;sd(X)>0} rearraging above equations to

\deqn{\mu = \frac{\mu^2}{\sigma^2}}

\deqn{\sigma = \frac{\mu}{\sigma^2}}

is possible such that the shape parameter $\alpha$
and rate parameter $\beta$ of the Gamma distribution are
obtained.
}
\examples{
# obtain shape and rate for Gamma distribution with mean = 1 and sd = 10
m = 1; s = 10
gam_pars = gamprior_repar(m_t = m, s_t = s)
shape = gam_pars[1] # shape parameter
rate = gam_pars[2] # rate parameter

# test: sample from the Gamma distribution with the obtained parameters
gamma_sample = rgamma(10000000, shape = shape, rate = rate)
m_emp = mean(gamma_sample); m_emp # estimated mean
s_emp = sd(gamma_sample); s_emp # estimated sd

}
