% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pgw.R
\name{plot_pgw}
\alias{plot_pgw}
\title{Visualize the power generalized Weibull distribution}
\usage{
plot_pgw(scale = 1, shape = 1, powershape = 1, ...)
}
\arguments{
\item{scale}{scale parameter}

\item{shape}{shape parameter}

\item{powershape}{power shape parameter}

\item{...}{optional plot arguments (with exceptions, eg. \code{type, lwd, main})}
}
\value{
A plot of the PDF, CDF, hazard and survival function
}
\description{
Visualize the power generalized Weibull distribution
}
\examples{
plot_pgw(scale = 2, shape = 5, powershape = 10)


}
\seealso{
\code{\link{pgw}}

An interactive version of this plot is available
on \url{https://janoleko.shinyapps.io/pgwd/}.
}
