% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.load.scenario.R
\name{sim.load.scenario}
\alias{sim.load.scenario}
\title{Load simulated scenario}
\usage{
sim.load.scenario(wd, pc, batchnr = 1, bayes = T)
}
\arguments{
\item{wd}{working directory where ALL the simulation files lie}

\item{pc}{parameter combination (numeric version)}

\item{batchnr}{nr of files (batches) containing simulation output based on the
same pc}

\item{bayes}{logical argument to specify whether to load the simulation results
obtained with Bayesian (, if \code{bayes = T}; default) or frequentist estimation
approach}
}
\value{
A dataframe \code{res.batch}.
}
\description{
Load all simulations of a simulated scenario (obtained using
\code{\link{sim.repeat.1.scenario}}).
}
