% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.merge_results.R
\name{sim.merge_results}
\alias{sim.merge_results}
\title{Merge result table batches from simulation study}
\usage{
sim.merge_results(pc_list, save = T, bayes = T)
}
\arguments{
\item{pc_list}{list of parameter combinations obtained from \link{sim.setup_simpars}}

\item{save}{if \code{TRUE} (default), merged table is saved as res_b.RData or res_f.RData
in same path where batches are stored;
else, it is returned to global environment}

\item{bayes}{\code{TRUE} (default), results of Bayesian simulations are merged,
else, results of frequentist simulations are merged}
}
\value{
Dataframe containing all simulation results (one repetition of one
simulation scenario per row). The
simulation parameters are stored in the first 9 columns. The remaining columns
contain
\enumerate{
\item posterior summary statistics and percentiles (ie information on the
posterior distribution) for each shape parameter and
\item posterior credibility
intervals (as specified in \code{$test} list element obtained from \link{sim.setup_simpars}).
}
}
\description{
Merges result table batches from simulation study obtained from using
\code{\link{sim.run}} or \code{\link{sim.run_parallel}}.
}
