% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.repeat.1.scenario.R
\name{sim.repeat.1.scenario}
\alias{sim.repeat.1.scenario}
\title{repeat simulation for one sample scenario (pc) in simulationstudy}
\usage{
sim.repeat.1.scenario(pc, pc_list, batch.ind)
}
\arguments{
\item{pc}{vector containing the parameter combination for data generation and model fitting, namely:
\itemize{
\item N: sample size
\item br: background rate
\item adr.rate: adverse event rate
\item adr.when: adverse event time
\item adr.relsd: relative standard deviation of the adverse event time
\item study.period: study period
\item tte.dist: time-to-event distribution
\item prior.dist: prior distribution
\item prior.belief: prior belief
}}

\item{pc_list}{list containing all parameter combinations for the simulation study
(necessary to extract further parameters for simulation study specification)}

\item{batch.ind}{integer indicating the pc batch index (for saving purposes)}
}
\value{
a data frame containing all fitting and posterior sample statistics of
the simulation runs (statistics are gathered using the
\code{\link{sim.stanfit.to.fitstats}} and \code{\link{sim.stanfit.to.poststats}}
functions).
}
\description{
Repeat the data generation and fitting and results output for
one scenario parameter combination (pc).
}
