% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.run.R
\name{sim.run}
\alias{sim.run}
\title{run all simulation scenarios}
\usage{
sim.run(pc_list)
}
\arguments{
\item{pc_list}{list of parameter combinations obtained from \link{\code{sim.setup_sim_pars}}

Runs simulations for all data generating processes, model and test alternatives
specified in pc_list. If part of the simulations is already run and saved in the
detemined resultpath, only missing simulations are run and saved.}
}
\description{
run all simulation scenarios
}
