% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.run_parallel.R
\name{sim.run_parallel}
\alias{sim.run_parallel}
\title{Run simulation (in parallel)}
\usage{
sim.run_parallel(pc_list, subset_ind = NULL)
}
\arguments{
\item{pc_list}{list of parameter combinations obtained from \code{\link{sim.setup_sim_pars}}}

\item{subset_ind}{vector of integers specifying which rows of \code{pc_list$pc_table}
to be considered in simulation runs}
}
\description{
Runs simulations for all data generating processes, model and test alternatives
specified in pc_list. If part of the simulations is already run and saved in the
detemined resultpath, only missing simulations are run and saved.
}
\details{
Simulation runs can be parallelized using the \code{\link[future]{future::plan}}
command (see example).
}
\examples{
\dontrun{
 # install.packages(future)
 future::plan(multisession, workers = availableCores()) # or another plan strategy
 sim.run_parallel(pc_list, subset_ind = NULL) # run all simulations
 
 
 # To run only a subset of simulation scenarios, specify the indices of the rows
 # in pc_list$pc_table that you want to run. For example, to run the first 10 scenarios:
 
 # install.packages(future)
 future::plan(multisession, workers = availableCores()) # or another plan strategy
 sim.run_parallel(pc_list, subset_ind = 1:10) # run first 10 simulation scenarios
}




}
