% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdat.R
\docType{data}
\name{testdat}
\alias{testdat}
\title{test data for development of simstudy code}
\format{
An object of class \code{data.frame} with 100 rows and 2 columns.
}
\usage{
testdat
}
\description{
generated with \code{datagen_tte(100, 0.1, 1, 0.5, 0.05, 365)}
command with N = 100, br = 0.1, adr.rate = 1, adr.when = 0.5, adr.relsd = 0.05, censor = 365.
}
\keyword{datasets}
