% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte2standat.R
\name{tte2standat}
\alias{tte2standat}
\title{reformat time-to-event data for usage in stanmodel.pgw}
\usage{
tte2standat(
  dat,
  scale.mean,
  scale.sd,
  shape.mean,
  shape.sd,
  powershape.mean,
  powershape.sd
)
}
\description{
Reformating of a matrix or dataframe containing a time-to-event
dataset with time in the first column and the event status in the second column.
}
\details{
Converts time-event data to a suitable format for rstan model fitting.
}
