% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte2standat.R
\name{tte2standat}
\alias{tte2standat}
\title{reformat time-to-event data for usage in \code{fit_pgw_tte()}}
\usage{
tte2standat(
  dat,
  mod = c("w", "pgw"),
  scale.mean = 1,
  scale.sd = 10,
  shape.mean = 1,
  shape.sd = 10,
  powershape.mean = 1,
  powershape.sd = 10
)
}
\arguments{
\item{dat}{Matrix or data frame with time in the first column and event status in the second column.}

\item{mod}{modelling approach to be used; options are "w" for Weibull or "pgw" for power generalized Weibull.}

\item{scale.mean}{The a priori expected mean of the scale parameter.}

\item{scale.sd}{The priori expected standard deviation of the scale parameter.}

\item{shape.mean}{The priori expected mean of the shape parameter.}

\item{shape.sd}{The priori expected standard deviation of the shape parameter.}

\item{powershape.mean}{The priori expected mean of the power shape parameter.}

\item{powershape.sd}{The priori expected standard deviation of the power shape parameter.}
}
\description{
Reformatting of a matrix or data frame containing a time-to-event
data set with time in the first column and the event status in the second column.
}
\details{
Converts time-event data to a suitable format for \code{rstan} model fitting.
}
